requirements.txt
openai>=1.12.0
anthropic>=0.18.0
transformers>=4.38.0
torch>=2.1.0
torchvision>=0.16.0
Pillow>=10.2.0
opencv-python>=4.9.0
pypdf>=4.0.0
pytesseract>=0.3.10
unstructured>=0.12.0
whisper>=1.1.10
moviepy>=1.0.3
chromadb>=0.4.22
langchain>=0.1.0
langchain-openai>=0.0.5
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
tqdm>=4.66.0
pydantic>=2.6.0
